09-serverless/homework/requirements.txt
onnxruntime
pillow
numpy
requests
